make the SVG occupy all the area of its parent div:
Added w-full h-full classes to the parent div to ensure it takes up all available space.
Changed the SVG's width and height attributes to "100%".
Added a viewBox attribute to the SVG that uses the computed width and height.
Modified the updateDimensions function to set both width and height based on the container's dimensions.
Added an event listener for the 'resize' event to update dimensions when the window is resized.
Added an onUnmounted hook to remove the event listener when the component is destroyed.
In the <style> section, set the .p-4 class to have width: 100% and height: 100%.